requirements.txt
aiohttp>=3.8.0
anthropic>=0.3.0
openai>=1.0.0
python-dotenv>=0.19.0
pydantic>=2.0.0
tweepy>=4.12.0
prompt_toolkit>=3.0.0
requests_oauthlib>=1.3.0 